Refactor python imports (#353)
Lack of order in imports leads to duplicates and makes it harder to understand the code.

The issue only refereced two files but with the isort utility every python file has been modified automatically.

The version of isort is mentioned in dev.txt. The black profile of isort is described in pyproject.toml and already used for this commit as default. The commands to run and check with isort have been added to the Makefile.